+++ example_command_results/android/app/src/main/kotlin/com/example/example_command_results/MainActivity.kt
@@ -1,5 +1,0 @@
-package com.example.example_command_results
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity()